Add files via upload
Object tracking based on HSV color space using the calcOpticalFlowFarneback() method. All the objects moving in a direction are displayed in the same color. The color of the objects depend on the direction of their movement. If objects are stationary, then they just appear mixed up in the dark background.